requirements.txt
fastapi
uvicorn
pydantic
python-multipart
PyJWT
python-dotenv
pytest
httpx
